move migration service to diagram service
