+++ .gitignore
@@ -148,13 +148,18 @@
 !td.vue/src/service/*.js
 !td.vue/src/service/api/
 !td.vue/src/service/api/*.js
+!td.vue/src/service/diagram/
+!td.vue/src/service/diagram/*.js
+!td.vue/src/service/diagram/v1/
+!td.vue/src/service/diagram/v1/*.js
 !td.vue/src/service/demo/
 !td.vue/src/service/demo/*.js
 !td.vue/src/service/demo/*.json
 !td.vue/src/service/entity/
 !td.vue/src/service/entity/*.js
 !td.vue/src/service/migration/
-!td.vue/src/service/migration/*.js
+!td.vue/src/service/migration/tmBom/
+!td.vue/src/service/migration/tmBom/*.js
 !td.vue/src/service/otm/
 !td.vue/src/service/otm/*.js
 !td.vue/src/service/provider/
@@ -223,12 +228,17 @@
 !td.vue/tests/unit/service/*.spec.js
 !td.vue/tests/unit/service/api/
 !td.vue/tests/unit/service/api/*.spec.js
+!td.vue/tests/unit/service/diagram/
+!td.vue/tests/unit/service/diagram/*.spec.js
+!td.vue/tests/unit/service/diagram/v1/
+!td.vue/tests/unit/service/diagram/v1/*.spec.js
 !td.vue/tests/unit/service/demo/
 !td.vue/tests/unit/service/demo/*.spec.js
 !td.vue/tests/unit/service/entity/
 !td.vue/tests/unit/service/entity/*.spec.js
 !td.vue/tests/unit/service/migration/
-!td.vue/tests/unit/service/migration/*.spec.js
+!td.vue/tests/unit/service/migration/tmBom/
+!td.vue/tests/unit/service/migration/tmBom/*.spec.js
 !td.vue/tests/unit/service/otm/
 !td.vue/tests/unit/service/otm/*.spec.js
 !td.vue/tests/unit/service/provider/
