+++ td.vue/src/service/diagram/v1/labels.js
